requirements.txt
Flask==2.3.2
google-cloud-storage==2.9.0